Change internals of FILE: reuse the unget buffer field to access an extended attribute data structure (pimpl idiom). Idea taken from citrus. Much discussion with deraadt@, otto@, millert@...
This is the least disruptive way to extend FILE, since its size can't really
change without this being a flag day. So the size doesn't change.

Actual additions to the structure will come in separate steps, since this
change is nasty enough on its own.

Tests by otto@ and others, careful reading of code by otto@ and millert@.

This is definitely a major bump, and has been checked to not impact a
full ports build.