Revamp ARMv7 ramdisk and miniroot creation process.
Instead of running mkuboot during install, we call it on during
the release build, so that we can also ship bootable images
of the generic and the ramdisk kernel.

We can now build miniroots for imx, too.  The installer also
can recognize the SoC and makes decision based on it.

Use ext2fs for i.MX6 based devices.

Tar the bootloader files to save some space.

ok syl@