Add support for '*', local-as and neighbor-as for ext-community matching and setting. This allows rules like:     ext-community * * # delete any ext-community     ext-community ovs * # delete any ext-community of specified type     ext-community rt 1.2.3.4:* and     ext-community rt 65001:local-as     ext-community rt local-as:11111
Note: Sometimes the type of the ext-community is underspecified when using
wildchars or expands.  So 'ext-community rt *' or 'ext-community soo *' will
match for any of the 3 possible types (2-byte AS, 4-byte AS and IP address).
If local-as/neighbor-as is used as an expand of as-number like
    ext-community rt local-as:11111
then bgpd will default to the 4-byte AS type to encode the community.

OK benno@